+++ noodles-gff/src/record/attributes/field/tag.rs
@@ -2,7 +2,7 @@ use std::borrow::Cow;
 
 use bstr::BStr;
 
-use super::percent_decode;
+use crate::record::percent_decode;
 
 pub(crate) fn parse_tag(src: &[u8]) -> Cow<'_, BStr> {
     percent_decode(src)